backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
edge-tts>=6.1.9
moviepy>=1.0.3
pydub>=0.25.1
Pillow>=10.0.0
numpy>=1.24.0
aiofiles>=23.2.1
google-generativeai>=0.3.0
requests>=2.31.0
pyngrok>=7.0.0
